Lib

Description

Helper and Utility Functions

Documentation

data PDK

Available PDKs

Constructors

XH035
X-Fab XH035 350nm Process

XH018
X-Fab XH018 180nm Process

XT018
X-Fab XT018 180nm Process

data Device

Device Types

Constructors

NMOS
NMOS

PMOS
PMOS

fst3 :: (a, b, c) -> a

First of Triple

boolMask :: [String] -> [String] -> [Bool]

Create a boolean mask from a subset of column names

boolMask' :: [String] -> [String] -> Tensor

Create a boolean mask Tensor from a subset of column names

intIdx :: [String] -> [String] -> [Int]

Create Integer Index for a subset of column names

intIdx' :: [String] -> [String] -> Tensor

Create Integer Index Tensor for a subset of column names

round :: Int -> Tensor -> Tensor

Round Float Tensor to n digits

defaultHeaders :: [String]

Default Column Names for stored Tensors

loadTensor :: FilePath -> IO Tensor

Load a Pickled Tensor from file

saveTensor :: Tensor -> FilePath -> IO ()

Pickle a Tensor and Save to file

currentTimeStamp :: String -> IO String

Current Timestamp as formatted string

currentTimeStamp' :: IO String

Current Timestamp with default formatting: "%Y%m%d-%H%M%S"

createModelDir :: String -> String -> IO String

Create a model archive directory for the given pdk and device type

data Args

Command Line Arguments

Constructors

Args

Fields

pdk :: PDK

PDK

dev :: Device

NMOS | PMOS

dir :: FilePath

Path to Data

num :: Int

Number of Epochs

size :: Int

Batch Size