Run

Description

Module for running training

Documentation

trainStep :: Tensor -> Tensor -> OpNet -> Adam -> IO (OpNet, Adam, Tensor)

Run one Update Step

trainingEpoch :: ProgressBar s -> [Tensor] -> [Tensor] -> [Tensor] -> OpNet -> Adam -> IO (OpNet, Adam, Tensor)

Run through all Batches performing an update for each

validStep :: Tensor -> Tensor -> OpNet -> IO Tensor

Run one Update Step

validationEpoch :: ProgressBar s -> [Tensor] -> [Tensor] -> OpNet -> [Tensor] -> IO Tensor

Run through all Batches performing an update for each

runEpochs :: FilePath -> Int -> Int -> Tensor -> Tensor -> Tensor -> Tensor -> OpNet -> Adam -> IO (OpNet, Adam)

Run Training and Validation for a given number of Epochs

run :: Args -> IO ()

Initiate Training Run for given Args

runExp :: Args -> IO ()

Experimental Training Run

runBase :: Args -> IO ()

Basic Training Run

testModel :: Int -> DataFrame Tensor -> [String] -> [String] -> (Tensor -> Tensor) -> IO ()

Model Test

testModelExp :: DataFrame Tensor -> [String] -> [String] -> (Tensor -> Tensor) -> IO ()

Experimental Model Test

region :: Int -> DataFrame Tensor -> Tensor

Filter region Experimental: -1 Saturation: 2 Sub-Threshold: 3