Lib

Description

Helper and Utility Functions

Documentation

data PDK

Available PDKs

Constructors

XH035
X-Fab XH035 350nm Process

XH018
X-Fab XH018 180nm Process

XT018
X-Fab XT018 180nm Process

GPDK180
Fictional Cadence GPDK180 GPDK 180nm Process

GPDK90
Fictional Cadence GPDK90 GPDK 90nm Process

SKY130
SkyWater SKY130 130nm Process

PTM130
Fictional Predictive Technology Models PTM130 130nm Process

data Device

Device Types

Constructors

NMOS
NMOS

PMOS
PMOS

fst3 :: (a, b, c) -> a

First of Triple

uncurry' :: (a -> b -> c -> d) -> (a, b, c) -> d

Uncurry triple

boolMask :: [String] -> [String] -> [Bool]

Create a boolean mask from a subset of column names

boolMask' :: [String] -> [String] -> Tensor

Create a boolean mask Tensor from a subset of column names

intIdx :: [String] -> [String] -> [Int]

Create Integer Index for a subset of column names

intIdx' :: [String] -> [String] -> Tensor

Create Integer Index Tensor for a subset of column names

round :: Int -> Tensor -> Tensor

Round Float Tensor to n digits

scale :: Tensor -> Tensor -> Tensor -> Tensor

Scale data to [0;1]

scale' :: Tensor -> Tensor -> Tensor -> Tensor

Un-Scale data from [0;1]

trafo :: Tensor -> Tensor -> Tensor

Apply log10 to masked data

trafo' :: Tensor -> Tensor -> Tensor

Apply pow10 to masked data

columnHeader :: PDK -> [String]

Default Column Names for stored Tensors

loadTensor :: FilePath -> IO Tensor

Load a Pickled Tensor from file

saveTensor :: Tensor -> FilePath -> IO ()

Pickle a Tensor and Save to file

currentTimeStamp :: String -> IO String

Current Timestamp as formatted string

currentTimeStamp' :: IO String

Current Timestamp with default formatting: "%Y%m%d-%H%M%S"

createModelDir :: String -> String -> IO String

Create a model archive directory for the given pdk and device type

data Args

Command Line Arguments

Constructors

Args

Fields

pdk :: PDK

PDK

dev :: Device

NMOS | PMOS

dir :: FilePath

Path to Data

num :: Int

Number of Epochs

size :: Int

Batch Size

trainStyle :: Int -> Style s

Progress Bar Style for Training

validStyle :: Int -> Style s

Progress Bar Style for Validation