Run

Description

Module for running training

Documentation

satMask :: Device -> DataFrame Tensor -> Tensor

Filter Datapoints not in saturation

process :: Tensor -> Tensor -> Tensor -> Tensor -> Tensor

Process Data

trainStep :: Tensor -> Tensor -> OpNet -> Adam -> IO (OpNet, Adam, Tensor)

Run one Update Step

trainingEpoch :: ProgressBar s -> [Tensor] -> [Tensor] -> [Tensor] -> OpNet -> Adam -> IO (OpNet, Adam, Tensor)

Run through all Batches performing an update for each

validStep :: Tensor -> Tensor -> OpNet -> IO Tensor

Run one Update Step

validationEpoch :: ProgressBar s -> [Tensor] -> [Tensor] -> OpNet -> [Tensor] -> IO Tensor

Run through all Batches performing an update for each

runEpochs :: FilePath -> Int -> [Tensor] -> [Tensor] -> [Tensor] -> [Tensor] -> OpNet -> Adam -> IO (OpNet, Adam)

Run Training and Validation for a given number of Epochs

run :: Args -> IO ()

Initiate Training Run for given Args